64bit: Fix compiler warnings in regard to 64bit
vty_interface_layer3.c:584:4: warning: format '%d' expects argument of type 'int', but argument 3 has type 'long unsigned int' [-Wformat=]
    sizeof(subscr->extension)-1, VTY_NEWLINE);

+++ openbsc/tests/gsm0408/gsm0408_test.c
@@ -48,7 +48,7 @@
 #define VERIFY(res, cmp, wanted)					\
 	if (!(res cmp wanted)) {					\
 		printf("ASSERT failed: %s:%d Wanted: %d %s %d\n",	\
-			__FILE__, __LINE__, res, # cmp, wanted);	\
+			__FILE__, __LINE__, (int) res, # cmp, (int) wanted);	\
 	}
